rename to northbound, southbound and message
