+++ eas-training/src/main/resources/mapper/CandidateMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.acl.port.repository.CandidateRepository" >
+<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.southbound.port.repository.CandidateRepository" >
     <resultMap id="candidateResult" type="Candidate" >
         <id column="id" property="id" jdbcType="VARCHAR"/>
         <result column="employeeId" property="employeeId" jdbcType="VARCHAR" />
