+++ eas-training/src/main/resources/mapper/MailTemplateMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.acl.port.repository.MailTemplateRepository" >
+<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.southbound.port.repository.MailTemplateRepository" >
     <resultMap id="mailTemplateResult" type="MailTemplate" >
         <id column="id" property="id" jdbcType="VARCHAR"/>
         <result column="template" property="template" jdbcType="VARCHAR" />
