+++ eas-training/src/main/resources/mapper/TicketHistoryMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.acl.port.repository.TicketHistoryRepository" >
+<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.southbound.port.repository.TicketHistoryRepository" >
     <resultMap id="ticketHistoryResult" type="TicketHistory" >
         <id column="id" property="id" jdbcType="VARCHAR"/>
         <result column="ticketId" property="ticketId.value" jdbcType="VARCHAR" />
