+++ eas-training/src/main/resources/mapper/TrainingMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.acl.port.repository.TrainingRepository" >
+<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.southbound.port.repository.TrainingRepository" >
     <resultMap id="trainingResult" type="Training" >
         <id column="id" property="id.value" jdbcType="VARCHAR"/>
         <result column="title" property="title" jdbcType="VARCHAR" />
