+++ eas-training/src/main/resources/mapper/ValidDateMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd" >
-<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.acl.port.repository.ValidDateRepository" >
+<mapper namespace="xyz.zhangyi.ddd.eas.valueadded.trainingcontext.southbound.port.repository.ValidDateRepository" >
     <resultMap id="validDateResult" type="ValidDate" >
         <id column="id" property="id" jdbcType="VARCHAR"/>
         <result column="trainingId" property="trainingId.value" jdbcType="VARCHAR" />
