+++ eas-training/src/test/java/xyz/zhangyi/ddd/eas/valueadded/trainingcontext/domain/training/TrainingServiceTest.java
@@ -1,7 +1,7 @@
 package xyz.zhangyi.ddd.eas.valueadded.trainingcontext.domain.training;
 
 import org.junit.Test;
-import xyz.zhangyi.ddd.eas.valueadded.trainingcontext.acl.port.repository.TrainingRepository;
+import xyz.zhangyi.ddd.eas.valueadded.trainingcontext.southbound.port.repository.TrainingRepository;
 import xyz.zhangyi.ddd.eas.valueadded.trainingcontext.domain.course.CourseId;
 
 import java.time.LocalDateTime;
